refactor: move tstype.py to units submodule
